Update README.md
firewall uses for allowing what comimg in and what goes out in a traffic, i demonstarted using two machine, kali window machine which IP is 192.168.1.19 and window IP is 192.168.1.18. the window can ping the kali but kali cannot be reached on the window, i allow my firewall but kali still cant reach the window.